Added Reconnect Status screen.  During WiT_RETRY, the Reconnect Status screen is display.  If Reconnection is successful (WiT_RECONNECT) then Locos and Consists are restored, ED returns to the previous Activity, the view is refreshed, Webviews reloaded to their prior state, etc.
Since Android does not reliably report Socket status, the only way to tell when the connection is lost is by attempting to send.  We do that at an outbound heartbeat rate that is calculated from the Heartbeat.  To ensure detection of a lost connection, I changed the behavior so that ED uses a default outbound heartbeat rate when the WiT heartbeat is disabled.  Since we currently allow WiT 4 seconds to respond, the time to detect a loss of connection can be long.  

Android does not seem very fast to reconnect to a network after losing the connection.  So although the reconnection process is automatic now, the time to reconnect to WiT seems quite a bit longer than desired.