Improved reconnect behavior.  Decreased TCP socket connect timeout from 3 seconds to 1 second to speed up the reconnect attempt rate - needs testing to confirm that is enough time to get a socket open on all devices.
Fixed WiT heartbeat so that both inbound and outbound heartbeats stop reliably during shutdown.  Corrected position of removeNotification in onResume for all Activities to ensure proper shutdown.  Filter which messages go to Activities when shutting down to speed things up.

Screens with webViews now load a static page in onPause and remove it in onResume to ensure that javascript doesn't continue to run when the screen is paused.