JmdnsImpl tolerates lack of joinGroup() support
The Jmdns implementation calls java.net.MulticastSocket.joinGroup() to
open a MulticastSocket.  The codes was crashing when this method was
missing from an Android 2.3.5 OS build.  I added a test for the presence
of this method in JmdnsImpl, via reflection, and simply omit the
joinGroup call when the method is missing.