Correct function button release handling for consists
The button press for functions was handled correctly, with the press
sent to all locos in the consist except for leadOnly functions.  However
the button release for functions was only being sent to the Lead loco
regardless of whether or not the function was leadOnly.  Corrected the
code for the release so it matches the code for the press - the release
is now sent to all locos unless the function is leadOnly.