Preserve throttle button states
Following a screen rotation, the Direction button states weren't
restored until an update is received from WiT.  Code now saves the
Requested Direction values in the bundle so the Direction button states
can be restored immediately.

Function button states are cleared when Menu or Status/Notification bar
is accessed.  Code now restores the Function button states when the Menu
or Status bar is exited and the Throttle screen gets the focus back.