ScreenLock handling fixes, Swipe/Fling direction selection
Corrected handling of start and end of gesture when the screen is locked
to fix a case where the screen could not be unlocked.

Changed gestureEnd so that left/right and up/down swipes are now
mutually exclusive. Swipes/flings are now processed for whichever
direction has the largest delta., which hopefully will suppress some
observed swipe anomalies.