Gamepad fixes, NACK tone on failed direction request
Gamepad dispatchKeyEvent method was "swallowing" keypresses that did not
pertain to ED.  Unfortunately this included the MENU button on older
devices.  Changed code so that all keypresses not in the gamePadKeys[]
are passed to the super for processing.

Changed code so repeat count is processed for each individual
gamePadKeys[] value rather than for all keys at once.  Similar to the
change above, this allows repeat to flow through for keys not in
gamePadKeys[].

changeDirectionIfAllowed() now returns a boolean indicating whether the
throttle direction matches the requested direction.  Gamepad uses this
return value to generate the "NACK" audio feedback tone on a throttle
direction that isn't honored (due to preference settings and current
speed).