CHANGED - Optional alternate labels for the throttle direction buttons NEW - long press on either direction button swaps the text but not the direction.
I did not get as far with this as I intended tonight. The button colours are not working properly yet.
This still uses a list of labels but I will change that to editable text tomorrow night.

The start of the code for checking the gamepad mode is there, but is not being called anywhere.