FIX - Immersive Mode (#217)
* NEW - Load transparent Splash Screen
Load a transparent splash screen on start, before loading the chosen theme.
Stops the checkerplate bitmap from momentarily being displayed when you are using any of the non-default themes.

* NEW - Change language from inside the App

* FIX - Splash screen on Old Android
The previous change to the splash screen will result in a transparent background on the connection and throttle pages on very old versions of android (e.g. 2.3.5)

* CHANGE - About text

* CHANGE - Some language corrections

* FIX - accelerometer required to install from the Play Store
no longer requires an accelerometer in the device to install from the Play Store (reported bug)

* FIX - Log Viewer backgound
The earlier change to the splash screen background resulted in the log view having a transparent background.

* FIX - Immersive Mode
ED would not return to immersive mode fully (part of the web view would show on the bottom of the screen) on start, and on return from a menu item.

* FIX - Immersive Mode
ED would not return to immersive mode fully (part of the web view would show on the bottom of the screen) on start, when you press back, but click no.