Added tests to prevent NPE (report 8)
It appears that throttle onResume/applySpeedRelatedOptions can run with
whichThrottle set to a consist that is null.  Not exactly sure of the
sequence that creates this (ESTOP may be involved) but added tests in
getConsist() and now return an empty Consist if the requested entry in
Consists[] is not available.  Hopefully this will address the NPE in
report 8.