Handle lack of activity.requestPermissions() in older Android version (report 1, pkg 2)
When a permission is not granted, ShowRetry called
requestNecessaryPermissions which uses activity.requestPermissions().
There was a @TargetApi annotation on this method.  A user ran ED on
Android 5.1 (Lollipop) got a NoSuchMethodError.  To accomodate this user
I modified ShowRetry to not call requestNecessaryPermissions on pre-6.0
(Marshmellow).  Also changed the notation to @RequiredApi so usage will
be flagged in red by AS.