avoid concurrentmodification
Change roster_entries, consist_entries and ConLoco.con from
LinkedHashMap to Collections.synchronizedMap(LinkedHashMap) to avoid
concurrency issues when copying the maps for reading.  Note that if this
works it would then be good to test whether we can eliminate the copying
and just do the reading on the original instances of the maps.