clearcache in webviews on start up
The throttle and web webviews cache downloaded javascript which can lead to using outdated code.  Added clearCache(true) in both OnCreate methods so that the javascript will be updated on the first load for any page.

Also added webview debugging capability so that timing, performance and code view can be performed remotely.