Initial fastclock value shown as 00:00
On a CONNECT message, the TA comm_handler sets fastClockSeconds to 0.  However the handler code was opening the WiT socket first and then later setting fastClockSeconds to 0.   As a result it was possible that the WiT Socket handler could receive the initial PFT time update and set fastClockSeconds first, and then the CONNECT handler would set it back to 0.  To prevent this case from happening I rearranged the CONNECT message handler code so that it sets fastClockSeconds before opening the WiTSocket.