revise restart on pref change
Implementing a great idea Peter had to change things so that when a Preference change causes a restart then CA relaunches CA and Throttle relaunches Throttle.  This means if you change a theme, etc, after you have already connected then now you will stay connected.

Followed Peter's approach and moved common methods to TA so that they would be accessible to both CA and Throttle.  getThrottleIntent() is used when (re)launching Throttle.  prefsForcedRestart() is used in CA and Throttle onCreate() to see if Preferences need to be shown.  Also moved the restart reason definitions to TA.