background mode handling
Cleaned up background mode handling.  Moved all code actions related to running in background mode to the TA ApplicationLifecycleHandler.  The app now runs the disconnect/shutdown sequence when it is about to get killed in background mode, which results in a clean environment for the next launch of the app.

Fixed string for inboundTimeoutInterval