FIX Preserve Web activity state - attempt 2 PLUS - fix for some swipe changes.
previously swipes to the activities, other than Throttle, started the target activity, but for Throttle it always just called finish()) which deleted the savedInstanceState.  Now it starts the Throttle instance like the other activities.
Note: onDestroy() still gets called so it is being removed from memory.

Changed to a common piece of code to calculate where to swipe to, to be more consistent.