PoM add repeat and convert CV to base 0
Programming on Main requires the DCC message to be sent twice in succession. the D1 command only sends the message once (repeat count 0 when tested via Loconet).  Changing to D2 sets the repeat count to 1.

The CV itself is 0-Base in the DCC Message, but 1-base in documents and the UI.  Changed to send (CV-1) in the DCC Message.