* added a new separate check for when the app is in background * reduce the amount of static text related to log messages (applicationName and activityName) * wholesale cleanup of the code to 'finish' an activity * additional log messages in select_loco
A particular tablet was going into background when the soft keyboard was being closed, and the system decided it need the memory and immediately killed ED.
The new 'check' is based around the activities telling t_a when that are in-view, or transitioning to another view.

Every call to log.d or log.e now use a static applicationName from t_a, and a static activityName from each class.
a) it will remove a lot of inconsistencies with the log messages
b) it should reduce the app size and memory usage a bit.